requirements.txt
mcstatus==4.0.0
python-dotenv==0.14.0